Split revision resolving from cloning
